Add support for a feature request in JIRA TK-2969 and JIRA TK-2970, where the user needed to be able to specify non standard "Decode As" options which cannot be stored in Wireshark's preferences file for certain older decoders such as FTP which do not support preferences very well.
pcap.rb:

Add logic for overriding default command line options, along with a
suffix which is appended to the user's options (if any), so that tshark
will always be provided with the mandatory options that are needed for
pcapr-local to work properly.

pcapr_local.rb:

Add bootstrap code that runs during pcapr.local's startup procedures
which fetches any custom command line options for tshark and overrides
the default options if the user specified different options.

config.rb:

Modify the default configuration template to include a blank set of
tshark options so that the user will have an empty field in a default
configuration file available for modification if and when needed.